+++ user-service/src/main/java/com/flight_booking/user_service/infrastructure/configuration/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.flight_booking.user_service.infrastructure.security;
+package com.flight_booking.user_service.infrastructure.configuration;
 
 import static org.springframework.security.config.http.SessionCreationPolicy.STATELESS;
 
@@ -13,6 +13,8 @@ import org.springframework.security.config.annotation.method.configuration.Enabl
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
@@ -33,6 +35,11 @@ public class SecurityConfig {
   @Bean
   public AuthenticationFilter authenticationFilter() {
     return new AuthenticationFilter(customUserDetailsService);
+  }
+
+  @Bean
+  public PasswordEncoder passwordEncoder() {
+    return new BCryptPasswordEncoder(10);
   }
 
   // SecurityFilterChain 설정
