--- user-service/src/main/java/com/flight_booking/user_service/infrastructure/security/jwt/PasswordEncoderConfig.java
@@ -1,18 +1,0 @@
-package com.flight_booking.user_service.infrastructure.security.jwt;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Component;
-
-@Component
-public class PasswordEncoderConfig {
-
-  private static final int STRENGTH = 10;
-
-  @Bean
-  public PasswordEncoder passwordEncoder() {
-    return new BCryptPasswordEncoder(STRENGTH);
-  }
-
-}
